refactor : 이벤트 함수 다른 폴더로 분리
- 멤버추가, 아이템 추가 등 모델을 바꾸는 이벤트 함수들 이벤트 폴더로 따로 분리하여 관리
